Totally rewrite the undo/redo functionality to support keeping track of the position of the selection box
The EventHistory class replaces CellHistory. Instead of storing sets of cells
which are changed, the history array stores events. Since actions can originate
from multiple objects (the select tool, the pencil tool, etc.), the originating
object is stored with each event. Anything that takes place that is undoable is
kept in an action object, which has three methods: `do`, `undo`, and `redo`. The
`do` method kicks off the event and is responsible for storing the "before" and
"after" states of whatever values are being set during the action so that the
`undo` and `redo` methods know what to roll back to. These values are kept in an
event object which is created on the fly and has a flexible (but consistent)
structure.